:recycle: refactor. UserCardBenefit ERD 변경에 따른 수정

+++ BE/rebirth/src/main/java/com/kkulmoo/rebirth/card/application/CardService.java
@@ -81,7 +81,7 @@ public class CardService {
             cardBenefits.add(
                     CardBenefit.builder()
                             .benefitCategory(categoryString)
-                            .receivedBenefitAmount((int) byUserIdAndBenefitId.getBenefitAmount())
+                            .receivedBenefitAmount(byUserIdAndBenefitId.getBenefitAmount())
                             .remainingBenefitAmount(
                                     calculateRemainingBenefit(
                                             benefitTemplate,
